fix deprecated method calls II

+++ Frameworks/BusinessLogic/ERAttachment/Sources/er/attachment/ERAttachmentRequestHandler.java
@@ -153,7 +153,7 @@ public class ERAttachmentRequestHandler extends WORequestHandler {
           }
 
           response.setStatus(200);
-          response.setContentStream(attachmentInputStream, bufferSize, (int) length);
+          response.setContentStream(attachmentInputStream, bufferSize, length);
         }
         catch (SecurityException e) {
           NSLog.out.appendln(e);
