+++ Frameworks/Core/ERExtensions/Sources/er/extensions/appserver/ERXKeepAliveResponse.java
@@ -70,7 +70,7 @@ public class ERXKeepAliveResponse extends ERXResponse {
 				}
 			}
 
-		}, 1, Integer.MAX_VALUE); // MS: turning it up to 11
+		}, 1, Long.MAX_VALUE); // MS: turning it up to 11
 	}
 
 	/**
